AssemblyParser: warn of calls to String.Format
Very basic detection of the pattern:

  String.Format(Catalog.GetString("{0} foo"), ...)

This is unsafe since an exception can be thrown
if a translator does not properly preserve format
arguments in the translated string.

Vernacular provides a safe wrapper API called
Catalog.Format which should be used instead.

In the future named/format arguments may be
supported directly in Catalog.GetString, but
this will do for now.