Treat empty "translated" values as untranslated
Before this patch, Vernacular assumed that if a translation
entry was provided at all it would be valid/translated. However,
it was possible to "provide" an empty translation, which should
mean "use the untranslated value." Without this patch, and in this
case, empty strings could be merged as translated values, which was
incorrect.

Essentially this fixes the case where a localized PO file has
entries such as:

  msgid "Untranslated default language string"
  msgstr ""

instead of:

  msgid "Untranslated default language string"
  msgstr "Untranslated default language string"

That is, both cases are now treated as identical.