Fix Android resource generation
When generating the resource files the strings are encoded to ensure correct XML entities are used. This was causing an
issue when strings that came from strings.xml (not from Catalog) had encoded entities. In this case the entities
would get double-encoded. This change unescapes any entities in source strings (whose origin is strings.xml), at resource
decode time, thus when the strings are encoded on resource generation they remain intact.